Add a favicon file
Most browsers try to download a /favicon.ico file or similar from the
web servers which is then e.g. used in the browser tabs or bookmark
lists. So let's add such a file for the QEMU website, too.

The files and HTML were created with realfavicongenerator.net from
QEMU's SVG-format logo.

Suggested-by: Thomas Huth <thuth@redhat.com>
Suggested-by: Daniel P. Berrange <berrange@redhat.com>
Message-Id: <1501075650-6971-1-git-send-email-thuth@redhat.com>
Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>