Link to git web view on contribute page
I was looking for a web view of qemu.git for reference, but
my first try was at https://www.qemu.org/contribute/ which
mentions how to clone but with no link for browsing.  Meanwhile,
the download page pointed people to the list of all repos (and
even then, only after you switch to the Source code tab), rather
than the qemu.git repo.  Make the link easier to find.

Signed-off-by: Eric Blake <eblake@redhat.com>