Add "Understanding QEMU devices" blog post
Last July, Eric Blake wrote a nice summary for newcomers about what
QEMU has to do to emulate devices for the guests. So far, we missed
integratating this somewhere into the QEM web site or wiki, so let's
publish this now as a nice blog post for the users.

Signed-off-by: Thomas Huth <thuth@redhat.com>
Message-Id: <1516958370-3955-1-git-send-email-thuth@redhat.com>

Incorporate editing suggestions made by others on the list,
particularly from Paolo Bonzini.

Signed-off-by: Eric Blake <eblake@redhat.com>
Message-Id: <20180209195436.975-1-eblake@redhat.com>
Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>