Add "Security Process" information to the main website
One reporter of a security issue recently complained that it might not
be the best idea to store our "Security Process" in the Wiki. Well, while
the page in the Wiki is protected (so that only some few people can edit
it), it is still possible that someone might find a bug in the Wiki
software to alter the page contents...
Anyway, it looks more trustworthy if we present the "Security Process"
information in the static website instead. Thus this patch adds the
information from the wiki to the Jekyll-based website now.

Signed-off-by: Thomas Huth <thuth@redhat.com>