blog post: Anatomy of a Boot, a QEMU perspective
While working on some test cases I realised there was quite a lot of
assumed knowledge about how things boot up. I thought it would be
worth gathering this together in a user facing document where we could
pour in the details and background to the boot process. The consensus
seems to be this is better as a blog post.

The post follows the socratic method and leaves the reader to ask
themselves some questions in an effort to elucidate them about any
problems they may be having before telling them that they probably
wanted to use virt-manager all along.

Based on Message-Id: <20200701161153.30988-2-alex.bennee@linaro.org>
Signed-off-by: Alex Bennée <alex.bennee@linaro.org>
[rth: review of doc form]
Reviewed-by: Richard Henderson <richard.henderson@linaro.org>
Message-Id: <20200703173948.22156-1-alex.bennee@linaro.org>
Signed-off-by: Thomas Huth <thuth@redhat.com>