Make page header nav narrower
When the browser window is narrow, but not yet switched into the mobile
layout, the page header nav will line wrap. This breaks layout
assumptions resulting in overlapping/obscured text.

This deals with the probem by reducing padding between the links, and
shortening "Documentation" to just "Docs".

This avoids the line wrapping, as well as making room for new links in
the future.

Signed-off-by: Daniel P. Berrangé <berrange@redhat.com>
Message-Id: <20201027132015.621733-5-berrange@redhat.com>
Signed-off-by: Thomas Huth <thuth@redhat.com>