Show only the latest three releases on the website
The QEMU project is only maintaining the very latest releases, so this
could be the current RC version, followed by the current stable release
and maybe also still the previous stable release (in case there are
severe bugs). But it does not make sense to show a fourth release that
we very likely do not update anymore. Thus don't show the fourth release
anymore to avoid that people still get the impression that we support
it (and then report bugs in that old release that we do not maintain
anymore).

Message-Id: <20210616110720.880586-2-thuth@redhat.com>
Signed-off-by: Thomas Huth <thuth@redhat.com>