Present the source code on the first tab of the download page
The QEMU project provides the source code of QEMU, and not any
binaries. So most people will come here for downloading the latest
version of the source code and not for getting instructions on
how to install the pre-packaged QEMU of their favourite distribution.
Thus let's put the information about the source code first.

Message-Id: <20210616110720.880586-5-thuth@redhat.com>
Signed-off-by: Thomas Huth <thuth@redhat.com>