Simplify and restructure the page footer
The footer is not an effective place to locate frequently needed site
navigation links. Some of the links are already present in the top
navigation bar, while others are just one jump away on the subsequent
page in a fairly obvious place.

This removes all the local site navigation links from the footer and
leaves it as a place for external project links, along with the website
license and SFC information.

Signed-off-by: Daniel P. BerrangÃ© <berrange@redhat.com>
Message-Id: <20210908122814.707744-6-berrange@redhat.com>
Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>