contribute: update hints for switch to the GitLab issue tracker
Currently, the bug reporting page has a paragraph about not sending patches
on the bug tracker, with a link to the patch submission guidelines.
Now that bugs are traceked on Gitlab it is more likely that prospective
contributors will submit a merge request linked to an issue, rather than
attaching a patch to the issue.  Update the language and, since it is
not limited to bug reports, add it to the main contribute page as well.

Also, "private" bugs are called "confidential" on the new tracker.

Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>