Add public key for tarball-signing to download page
We used to have public keys listed on the SecurityProcess page back
when it was still part of the wiki, but they are no longer available
there and some users have asked where to obtain them so they can verify
the tarball signatures.

That was probably not a great place for them anyway, so address this by
adding the public signing key directly to the download page.

Reported-by: Stefan Hajnoczi <stefanha@redhat.com>
Reviewed-by: Thomas Huth <thuth@redhat.com>
[mdr: incorporate formatting fixup/suggestions from Daniel]
Signed-off-by: Michael Roth <michael.roth@amd.com>