Add signing pubkey for python-qemu-qmp package
Add the pubkey currently used for signing PyPI releases of qemu.qmp to a
stable location where it can be referenced by e.g. Fedora RPM specfiles.

At present, the key happens to just simply be my own -- but future
releases may be signed by a different key. In that case, we can
increment '1.txt' to '2.txt' and so on. The old keys should be left in
place.

The format for the keyfile was chosen by copying what OpenStack was
doing:
https://releases.openstack.org/_static/0x2426b928085a020d8a90d0d879ab7008d0896c8a.txt

Generated with:
> gpg --with-fingerprint --list-keys jsnow@redhat.com > pubkey
> gpg --armor --export jsnow@redhat.com >> pubkey

Signed-off-by: John Snow <jsnow@redhat.com>
Message-Id: <20220818165358.217613-1-jsnow@redhat.com>
Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>