include Dockerfile used to proxy qemu.org to gitlab.io
The server configuration used to actually publish qemu-web.git
to qemu.org together with the sphinx documentation, right now,
is hidden in the machine that runs qemu.org.

Include it in the qemu-web.git website for documentation purpose,
in the format of a Dockerfile that you can use to run your own
qemu.org website.

Note: this container is not yet in use, but it is already live
and can be tested by adding

20.218.253.93   qemu.org www.qemu.org git.qemu.org qemu-project.org www.qemu-project.org git.qemu-project.org

to your /etc/hosts file.

Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>