Give re-uploaded 8.2.3 tarballs a new name
The 8.2.3 tarballs needed to be re-uploaded shortly after release, but
the CDN seems to still be serving the old copies in some cases. In the
past it has taken upwards of a month for the CDN to refresh certain
files/page listings so to avoid continued issues until then go ahead and
rename the tarballs to force the update. The added "-reupload" isn't too
dissimilar to the "-rcX" tags so is hopefully not too disruptive to
downstreams.

In the future it might be better to just bump/retag the release when
such a thing is necessary.

Signed-off-by: Michael Roth <michael.roth@amd.com>