sponsors: split off paypal donations info to a separate page
The link to the paypal donations page is hidden in the middle of
the sponsorship page, which makes it unlikely the a casual reader
will notice it.

Since ongoing formal sponsors are a quite different arrangement
from adhoc donations, both in terms of scale and whom the target
audience is for each, they deserve separate pages.

This will facilitate linking directly to the donation information
from other places on the website, without distraction from the
corporate sponshorship information.

Signed-off-by: Daniel P. Berrangé <berrange@redhat.com>