download: don't suggest running 'git submodule'
The QEMU build system will automatically trigger updates of any git
submodules that are required at the correct time. There is no reason
for users to manually update submodules after checkout. This is a big
waste of time since some of the submodules are huge and will never be
touched. The suggestion to recursively update them also hit users
when an EDK2 submodule recently went missing.

Signed-off-by: Daniel P. Berrangé <berrange@redhat.com>