download: Limit download page to actively-supported releases
8.0.x and 8.1.x are no longer actively maintained and each nearing at
least a year since the last update, so go ahead and remove them to make
room for active releases.

Currently 7.2/8.2 are LTS releases, and 9.0/9.1/9.2 are non-LTS but
still have recent or future releases associated. So also set the display
limit to 5 so only these show up on the download page.

Signed-off-by: Michael Roth <michael.roth@amd.com>