security: drop info about GPG mail to secalert@redhat.com
We want the qemu-security mailing list to be considered the
authoritative point of contact for all security bugs. The
goal is try to ensure that upstream has visibility on all
security bugs and things don't lost by vendors who might
close an issue as not affecting their product configuration
and accidentally forget to forward it upstream.

The desire to use GPG for security reports isn't common
enough in QEMU context to justify pointing people to the
Red Hat secalert address.

Signed-off-by: Daniel P. Berrangé <berrange@redhat.com>